Update Sample HTML Basic Web Page.txt
This is the sample format for you to create a website in IPFS. This is one of the most basic websites and won't win any awards, but hopefully it will show you how to work with IPFS. Simply download (or copy/paste the content) the file into a txt editor and change out my social media links for your own. Don't forget to save the file as .html before dragging to ipfs. Remember that IPFS can take more than 10 minutes to process your page.